fix(block_graph): `merge_chains` errors if no point of agreement
Previously BlockGraph allowed merging on an unrooted checkpoint
if it only extended the tip. This is too permissive because the
CP itself gives no indication of where it is supposed to
connect.

Since the design of BlockGraph requires every block to declare a
parent, we disallow connecting a single CP unless there is an
unambiguous point of connection.